requirements.txt
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.2
langgraph>=0.0.20
openai>=1.3.0
chromadb>=0.4.18
pydantic>=2.5.0
langchain-community>=0.0.10
pypdf>=3.15.1
tiktoken>=0.5.1
